Obtain the PIT-values of a model.

A function to obtain the (possibly uncalibrated) PIT-values of any fitted model that assumes a normal distribution for the output, such as (but not limited to), a lm() or a neural network that used the Mean Squared Error as the loss function.

PIT_global(ycal, yhat, mse)

Arguments

ycal

observations of the recalibration set

yhat

predictions of the recalibration set from the uncalibrated model

mse

Mean Squared Error of validation set.

Value

Vector of PIT-values

Examples

n <- 10000
split <- 0.8

# generating heterocedastic data
mu <- function(x1){
10 + 5*x1^2
}

sigma_v <- function(x1){
30*x1
}

x <- runif(n, 1, 10)
y <- rnorm(n, mu(x), sigma_v(x))

x_train <- x[1:(n*split)]
y_train <- y[1:(n*split)]

x_cal <- x[(n*split+1):n]
y_cal <- y[(n*split+1):n]

model <- lm(y_train ~ x_train)

y_hat <- predict(model, newdata=data.frame(x_train=x_cal))

MSE_cal <- mean((y_hat - y_cal)^2)

PIT_global(ycal=y_cal, yhat=y_hat, mse=MSE_cal)
#>    [1] 3.209528e-01 2.542822e-01 5.848373e-01 5.522800e-01 4.941958e-01
#>    [6] 2.841665e-01 4.765463e-01 5.241583e-01 4.617327e-01 6.711078e-01
#>   [11] 4.987337e-01 9.814634e-01 6.186309e-01 7.613117e-01 9.443548e-01
#>   [16] 1.639562e-01 4.994037e-01 5.379369e-01 4.949492e-01 5.750353e-01
#>   [21] 7.634381e-01 2.578784e-01 7.400041e-01 6.948636e-01 2.356780e-01
#>   [26] 1.426774e-01 3.843668e-01 5.461879e-01 5.221393e-01 6.838410e-01
#>   [31] 2.170843e-01 2.533476e-01 4.953253e-01 6.200301e-01 2.582158e-01
#>   [36] 5.597634e-01 6.762802e-01 9.554940e-01 5.049130e-04 4.145003e-01
#>   [41] 4.177099e-01 9.110754e-01 2.208708e-02 6.678141e-01 3.025740e-01
#>   [46] 1.699051e-02 4.342269e-03 7.429606e-01 9.814865e-01 8.575487e-01
#>   [51] 9.387832e-01 6.878286e-01 1.687177e-01 6.729524e-01 6.847628e-01
#>   [56] 4.265289e-01 5.879793e-02 5.246961e-01 6.578146e-01 6.057525e-01
#>   [61] 4.365575e-01 5.393396e-01 6.937027e-01 8.371172e-01 7.577937e-01
#>   [66] 5.828714e-01 8.861078e-01 6.996180e-01 6.012944e-02 4.528861e-01
#>   [71] 8.683483e-01 6.473983e-01 4.278458e-01 7.266505e-01 3.906435e-01
#>   [76] 1.872083e-01 3.380191e-01 7.389251e-02 1.215886e-01 5.472933e-01
#>   [81] 5.085622e-01 3.038262e-01 2.446219e-01 9.095126e-01 3.168781e-01
#>   [86] 9.328806e-01 9.664716e-01 1.494852e-01 5.576187e-01 4.687967e-01
#>   [91] 2.684255e-01 3.822580e-01 4.811633e-01 3.373678e-01 7.426035e-01
#>   [96] 4.506117e-01 3.326797e-01 5.582804e-01 9.471109e-01 3.850320e-02
#>  [101] 3.566103e-01 5.078086e-01 2.648426e-02 5.366813e-01 6.735404e-01
#>  [106] 6.060324e-01 8.660170e-01 7.944941e-01 4.069576e-01 7.964448e-01
#>  [111] 7.233757e-01 2.721450e-01 8.525928e-01 2.375411e-02 6.747939e-01
#>  [116] 3.505730e-01 5.079863e-01 8.324804e-01 2.236901e-02 4.009909e-01
#>  [121] 2.238197e-01 9.481683e-01 3.857298e-02 7.181686e-01 7.277364e-01
#>  [126] 3.446236e-01 3.065754e-01 5.349311e-01 3.094180e-01 5.704436e-01
#>  [131] 6.844809e-02 8.330340e-01 8.469795e-01 4.004108e-01 9.980528e-01
#>  [136] 5.540584e-01 9.516905e-01 5.546023e-02 9.701903e-01 7.056137e-01
#>  [141] 3.879256e-01 1.927386e-01 3.609027e-04 1.457337e-01 9.977579e-01
#>  [146] 4.629744e-01 2.989980e-01 3.408646e-01 6.084183e-01 2.773938e-01
#>  [151] 5.029694e-01 9.329656e-01 2.884244e-01 3.248259e-01 5.681085e-01
#>  [156] 8.586596e-01 7.771373e-01 7.117091e-01 8.566002e-02 3.038173e-01
#>  [161] 5.386503e-03 1.167573e-01 9.756200e-01 5.972398e-01 4.890154e-01
#>  [166] 1.204637e-01 2.802092e-01 3.526749e-01 4.428582e-01 5.923338e-01
#>  [171] 4.507230e-01 7.014757e-01 1.415888e-01 3.723933e-01 3.580957e-01
#>  [176] 3.336864e-01 4.964990e-01 3.226519e-01 9.830101e-01 6.422924e-01
#>  [181] 7.032791e-01 7.192926e-01 6.001618e-01 9.998183e-01 1.782125e-02
#>  [186] 3.001945e-01 3.626527e-01 5.746659e-01 3.955776e-01 7.610542e-01
#>  [191] 3.887457e-01 4.730374e-01 4.950890e-01 6.418405e-01 5.006575e-01
#>  [196] 1.557800e-01 6.473796e-01 9.799437e-01 9.574561e-01 8.712224e-01
#>  [201] 5.813606e-01 4.903663e-01 6.836948e-01 4.296365e-01 9.986650e-01
#>  [206] 5.492456e-01 7.225288e-01 6.670911e-01 2.264972e-01 8.846454e-01
#>  [211] 5.778253e-01 2.113359e-02 2.612052e-01 8.993121e-01 9.273065e-01
#>  [216] 2.379286e-01 6.021698e-01 3.544872e-01 1.564039e-01 4.246967e-01
#>  [221] 6.366790e-01 6.725571e-01 8.069578e-01 8.153741e-01 1.854190e-01
#>  [226] 2.172898e-01 1.684592e-01 7.676812e-01 9.970012e-01 6.569859e-01
#>  [231] 6.139197e-01 4.847418e-01 7.926957e-02 7.700962e-01 4.697670e-01
#>  [236] 8.656109e-01 5.696149e-01 8.717138e-01 7.574000e-01 9.264827e-01
#>  [241] 4.918852e-01 6.308719e-01 2.998063e-01 5.440481e-01 6.670092e-01
#>  [246] 2.267068e-01 9.582720e-05 8.601971e-01 3.490073e-01 5.364982e-01
#>  [251] 7.879091e-02 7.670277e-01 6.644234e-01 7.966409e-01 5.950357e-03
#>  [256] 8.038596e-01 7.726666e-01 4.161262e-01 6.980985e-01 2.355243e-01
#>  [261] 6.573912e-01 9.905392e-01 7.469150e-01 3.217261e-01 9.624244e-01
#>  [266] 4.321638e-01 5.911148e-01 7.789989e-01 3.779006e-01 6.548775e-01
#>  [271] 9.107512e-02 5.600097e-01 3.761029e-01 5.586919e-01 4.668294e-02
#>  [276] 1.486918e-01 3.050985e-01 9.972142e-01 9.049245e-01 1.932052e-01
#>  [281] 4.405279e-01 2.562198e-01 5.100040e-01 4.505432e-01 4.441820e-01
#>  [286] 1.636154e-01 7.289623e-03 8.586065e-02 5.807717e-01 7.050007e-02
#>  [291] 1.206648e-01 1.670050e-01 3.008639e-02 6.618554e-01 5.071989e-01
#>  [296] 4.476212e-01 2.105572e-01 2.950113e-01 9.718952e-01 6.584970e-01
#>  [301] 7.324243e-02 9.499303e-01 3.576326e-01 9.998428e-01 6.158061e-01
#>  [306] 5.961305e-01 3.432620e-01 4.919524e-01 8.172159e-01 2.246073e-01
#>  [311] 5.956185e-01 2.137342e-01 5.803162e-01 1.175724e-01 1.384600e-01
#>  [316] 9.804188e-01 4.811496e-01 6.181114e-01 9.499617e-02 5.378143e-01
#>  [321] 1.912286e-01 1.329244e-01 7.696357e-01 2.695591e-01 1.874722e-01
#>  [326] 1.101252e-01 6.079470e-02 9.442265e-01 5.938891e-01 2.420370e-01
#>  [331] 5.702224e-01 5.053205e-01 2.144031e-01 6.784454e-01 2.681311e-01
#>  [336] 6.694944e-01 7.017074e-01 7.247929e-01 3.040990e-01 8.602378e-01
#>  [341] 1.009735e-01 4.340843e-01 1.836635e-01 2.284394e-01 9.792969e-01
#>  [346] 6.274403e-01 4.581579e-01 3.584369e-01 7.040197e-01 5.019358e-01
#>  [351] 2.341444e-01 2.274056e-01 1.749485e-01 5.978402e-01 6.809092e-01
#>  [356] 9.994246e-01 9.997210e-01 9.585073e-01 9.828154e-01 5.503842e-01
#>  [361] 4.105724e-01 2.054716e-01 2.547407e-01 3.205622e-01 4.290653e-01
#>  [366] 4.647279e-01 8.954724e-01 8.583492e-01 2.187015e-01 3.884022e-01
#>  [371] 9.706440e-01 5.852712e-01 4.899380e-01 9.610651e-01 9.391375e-01
#>  [376] 4.385380e-01 7.865182e-02 7.064022e-01 7.911827e-01 9.084079e-01
#>  [381] 7.308023e-01 3.864131e-01 6.684720e-01 9.238401e-01 9.470780e-01
#>  [386] 5.103842e-01 6.168363e-01 5.290780e-01 2.480770e-01 5.501589e-03
#>  [391] 1.820832e-01 7.668523e-01 5.551558e-01 2.143299e-01 2.140224e-01
#>  [396] 2.748238e-01 4.571362e-01 7.687427e-01 9.234780e-01 6.128225e-01
#>  [401] 6.332560e-01 1.079530e-01 6.438584e-01 3.136279e-02 9.675936e-01
#>  [406] 1.007531e-01 7.737288e-01 8.739403e-01 9.371710e-01 9.972147e-01
#>  [411] 1.155770e-02 7.098008e-01 2.633398e-01 9.825497e-01 4.235528e-01
#>  [416] 6.142647e-01 8.039558e-01 3.030510e-01 9.514572e-01 6.891465e-01
#>  [421] 6.160320e-01 6.667731e-01 2.682230e-01 6.882917e-01 3.197736e-01
#>  [426] 4.118900e-01 8.619762e-01 3.506029e-02 2.983166e-01 9.679655e-01
#>  [431] 6.462806e-01 7.165370e-01 8.885618e-01 5.260310e-01 4.671826e-01
#>  [436] 9.536811e-01 7.236465e-01 4.576799e-01 8.349414e-01 5.634648e-01
#>  [441] 5.482312e-01 3.470745e-01 6.540868e-01 7.979455e-01 6.108402e-01
#>  [446] 7.189960e-01 4.901169e-01 5.330378e-01 3.444335e-01 2.217240e-01
#>  [451] 8.156262e-01 6.148483e-01 1.263090e-01 6.989804e-01 5.330237e-01
#>  [456] 9.976841e-01 2.834725e-01 9.852655e-01 4.176053e-01 3.251531e-01
#>  [461] 7.024314e-01 3.565683e-01 2.673699e-01 7.186028e-01 3.957805e-01
#>  [466] 7.169461e-01 3.940913e-01 7.070158e-01 5.635226e-01 7.710455e-01
#>  [471] 6.479522e-01 3.466959e-01 1.172514e-01 4.054321e-01 9.230608e-02
#>  [476] 5.263395e-01 1.666226e-01 5.470581e-01 5.297409e-01 6.538907e-01
#>  [481] 4.146060e-01 7.270154e-01 6.428663e-02 5.885199e-01 2.707255e-01
#>  [486] 8.544356e-02 2.375469e-01 2.094058e-01 5.681658e-01 2.549011e-01
#>  [491] 2.404293e-02 3.159765e-01 5.178230e-01 3.274941e-01 9.099284e-01
#>  [496] 6.199290e-01 6.978089e-02 8.414753e-01 7.930515e-01 3.488702e-01
#>  [501] 3.210221e-01 3.192779e-01 2.504349e-01 6.718586e-01 6.966701e-01
#>  [506] 7.640630e-01 9.765603e-01 7.787838e-02 3.494088e-01 6.230563e-01
#>  [511] 3.931817e-01 4.749640e-01 7.466248e-01 2.907655e-01 9.408910e-01
#>  [516] 2.825798e-01 6.432390e-01 3.557594e-01 6.454840e-01 1.857779e-01
#>  [521] 6.410991e-01 4.846269e-01 3.128575e-01 3.276353e-01 6.575424e-01
#>  [526] 2.964029e-01 6.770094e-01 6.859469e-01 7.581518e-01 3.386725e-01
#>  [531] 5.701291e-01 7.903824e-01 8.162299e-01 9.745326e-01 7.077387e-02
#>  [536] 7.938761e-01 4.229804e-01 8.328790e-01 9.277912e-01 6.948243e-01
#>  [541] 5.178250e-01 3.578075e-01 2.994095e-02 3.558267e-01 3.613691e-01
#>  [546] 7.083694e-01 2.287719e-01 5.703143e-01 5.262076e-02 4.700596e-01
#>  [551] 5.492869e-01 3.364416e-01 9.139081e-01 4.783198e-01 9.600436e-02
#>  [556] 4.014401e-01 6.715411e-01 6.110722e-01 8.955153e-01 2.625834e-01
#>  [561] 4.869918e-01 5.358802e-01 1.938409e-01 8.982452e-01 2.247116e-01
#>  [566] 6.916782e-01 7.677548e-01 4.690357e-01 5.442777e-01 1.174217e-01
#>  [571] 6.332673e-01 8.716147e-02 4.444707e-01 5.362228e-01 5.524600e-01
#>  [576] 4.160168e-01 4.651853e-01 5.639764e-01 8.292627e-01 7.750013e-01
#>  [581] 5.226119e-01 3.800509e-01 3.560815e-01 9.058208e-01 2.904927e-01
#>  [586] 1.819415e-01 7.790951e-01 3.024187e-01 5.787625e-01 8.920494e-01
#>  [591] 1.602084e-01 5.098210e-01 6.531631e-03 5.749816e-01 6.033913e-01
#>  [596] 8.731474e-01 5.682103e-01 6.889114e-01 4.986541e-01 5.659835e-01
#>  [601] 5.858882e-01 6.012657e-01 6.410998e-01 7.902950e-01 6.376906e-01
#>  [606] 5.527858e-01 9.964609e-01 4.436712e-01 4.392464e-01 4.455992e-01
#>  [611] 7.331340e-01 4.785280e-02 7.961712e-01 6.002810e-01 9.523661e-01
#>  [616] 5.928954e-01 5.685669e-01 4.444605e-01 8.282959e-02 6.321323e-01
#>  [621] 5.170730e-01 3.477551e-01 4.680966e-01 8.427991e-01 4.278012e-01
#>  [626] 5.241894e-01 8.358471e-01 3.892099e-01 8.961772e-01 2.992590e-01
#>  [631] 5.611537e-01 9.755196e-02 4.193524e-01 8.056997e-01 3.269908e-01
#>  [636] 4.265333e-01 9.226137e-01 6.053548e-01 9.628666e-01 2.094042e-01
#>  [641] 9.423321e-01 2.594233e-01 1.601184e-01 3.782354e-01 4.162561e-01
#>  [646] 3.628194e-01 5.673326e-01 3.620361e-01 1.256250e-02 5.237216e-01
#>  [651] 9.102740e-01 1.376892e-01 7.734860e-01 4.727682e-01 8.112410e-01
#>  [656] 7.300581e-01 2.146690e-01 9.996660e-01 5.994690e-01 4.887470e-01
#>  [661] 2.890383e-01 5.429336e-01 2.794233e-01 5.567439e-01 3.661317e-01
#>  [666] 3.625172e-01 3.528648e-01 9.458979e-01 1.966751e-01 6.035613e-01
#>  [671] 3.059878e-01 5.909469e-01 4.552182e-01 7.235039e-01 8.391237e-01
#>  [676] 5.188315e-01 3.732818e-01 7.056301e-01 4.985309e-01 7.821846e-01
#>  [681] 6.027216e-01 1.745843e-01 1.984140e-01 8.998867e-01 5.318924e-01
#>  [686] 5.850282e-01 7.865049e-01 4.224303e-01 3.151797e-01 4.125238e-01
#>  [691] 5.842601e-01 5.337467e-02 4.128630e-01 5.238856e-01 2.429520e-01
#>  [696] 1.650808e-01 5.850224e-01 3.559747e-02 3.277866e-01 1.957318e-01
#>  [701] 6.774947e-01 4.797203e-01 8.947257e-01 2.559951e-02 2.353754e-01
#>  [706] 5.486971e-02 7.792366e-01 8.890431e-01 6.375221e-01 5.617598e-01
#>  [711] 6.617409e-01 4.723919e-01 4.964649e-01 6.218685e-01 4.989846e-01
#>  [716] 5.016222e-01 5.851027e-01 4.612707e-01 5.836774e-01 3.627334e-01
#>  [721] 6.903366e-01 6.065970e-01 5.694647e-01 2.287998e-01 4.085515e-01
#>  [726] 5.857110e-01 5.833096e-01 2.565479e-01 5.929932e-01 5.851739e-02
#>  [731] 9.433526e-01 5.274391e-01 1.179636e-01 5.464003e-01 3.605678e-02
#>  [736] 9.038981e-01 2.536475e-04 5.689570e-01 6.891879e-01 9.943144e-01
#>  [741] 4.863624e-01 5.878660e-01 6.319555e-01 4.038334e-01 7.580725e-02
#>  [746] 5.542493e-01 5.527634e-01 6.658606e-01 6.241884e-01 4.037474e-01
#>  [751] 5.826828e-01 6.476843e-01 2.660224e-01 2.174323e-02 3.719546e-01
#>  [756] 3.524815e-01 1.521963e-01 4.078132e-01 2.029477e-01 8.006332e-03
#>  [761] 4.534977e-01 6.763963e-03 5.526746e-01 1.647759e-01 4.093808e-01
#>  [766] 7.321555e-01 4.956426e-01 6.801390e-01 1.281679e-01 3.191613e-01
#>  [771] 8.773691e-01 5.290820e-01 6.244100e-02 6.687448e-01 3.290061e-01
#>  [776] 1.205455e-01 9.998865e-01 6.278460e-02 2.823239e-01 5.248586e-01
#>  [781] 8.388841e-02 3.698869e-01 6.584592e-01 3.189524e-01 4.290356e-01
#>  [786] 6.076814e-01 7.198238e-01 4.588289e-01 5.230503e-01 5.717004e-01
#>  [791] 4.786573e-01 3.315674e-01 1.508549e-01 4.314851e-01 4.162416e-01
#>  [796] 7.768009e-01 5.550853e-01 7.196970e-01 5.693347e-01 7.171248e-01
#>  [801] 8.739105e-01 8.693778e-01 5.940658e-01 7.224398e-01 2.673465e-01
#>  [806] 5.219425e-01 4.309676e-01 5.599067e-01 8.773491e-01 7.274320e-01
#>  [811] 7.953347e-01 2.903813e-01 6.978498e-01 7.830920e-01 5.600471e-01
#>  [816] 9.599369e-01 4.472307e-02 5.094630e-01 9.781615e-01 2.761436e-01
#>  [821] 9.240027e-01 8.509987e-01 1.259784e-01 3.254186e-01 7.407602e-01
#>  [826] 6.840154e-01 2.753709e-01 7.105855e-01 6.755479e-01 4.318736e-01
#>  [831] 7.955869e-01 4.004440e-01 3.547024e-01 3.904591e-01 6.784898e-01
#>  [836] 2.130456e-01 5.658298e-01 6.957480e-01 2.214241e-01 2.871698e-01
#>  [841] 8.225068e-01 9.767322e-01 5.217971e-01 3.797647e-01 9.213658e-01
#>  [846] 3.886637e-01 3.890143e-01 3.863363e-01 4.200584e-01 7.741113e-01
#>  [851] 3.836614e-02 2.120880e-01 9.690115e-01 2.542654e-01 4.985964e-01
#>  [856] 5.458072e-01 2.112351e-03 2.372197e-01 6.728505e-01 8.428411e-01
#>  [861] 8.961575e-01 6.993551e-02 9.500087e-01 4.642248e-01 1.314812e-01
#>  [866] 7.706905e-01 6.253854e-01 5.073744e-02 6.797845e-01 4.505858e-02
#>  [871] 5.719593e-01 7.363615e-03 5.862395e-01 3.858178e-01 2.650789e-01
#>  [876] 3.360866e-02 1.346614e-01 9.813141e-01 4.364005e-01 9.978640e-01
#>  [881] 1.041081e-01 3.998935e-01 8.647256e-02 4.346847e-01 8.409433e-01
#>  [886] 5.609372e-01 1.055420e-01 5.940621e-01 1.687162e-01 1.244945e-01
#>  [891] 6.156444e-01 1.637213e-01 9.867550e-01 7.927801e-01 4.002800e-01
#>  [896] 3.561232e-01 4.333962e-01 6.258985e-01 5.525042e-01 4.181832e-01
#>  [901] 5.818901e-01 6.921128e-01 7.271299e-01 2.783615e-01 5.248492e-01
#>  [906] 4.884142e-01 7.256810e-01 8.439506e-01 7.040021e-01 7.046578e-01
#>  [911] 2.131645e-01 2.491841e-01 7.460119e-02 1.303402e-01 1.059486e-01
#>  [916] 3.921982e-01 4.733819e-01 1.887045e-01 8.281211e-01 4.044133e-01
#>  [921] 9.398445e-01 7.695054e-01 6.649416e-01 6.137524e-01 1.906489e-01
#>  [926] 6.677409e-01 5.465527e-01 4.785903e-01 3.378761e-01 8.966192e-01
#>  [931] 5.246453e-01 3.821887e-01 6.445853e-02 9.410329e-02 9.143277e-01
#>  [936] 1.999781e-01 1.573927e-01 5.617061e-01 5.303251e-01 7.034624e-01
#>  [941] 5.572455e-01 5.195402e-01 5.270747e-01 5.238411e-01 1.897732e-01
#>  [946] 5.329997e-01 8.189819e-01 2.410554e-01 5.963731e-01 9.684506e-01
#>  [951] 6.556025e-01 2.897049e-01 3.887374e-01 5.311379e-01 5.804273e-01
#>  [956] 8.936845e-01 7.282677e-01 1.612445e-01 3.593689e-01 4.968767e-01
#>  [961] 4.653106e-01 2.955793e-01 7.582425e-01 2.694780e-02 9.953257e-01
#>  [966] 2.162299e-01 8.849053e-01 1.075511e-01 7.694038e-01 3.148217e-01
#>  [971] 8.730147e-01 1.376396e-01 3.078361e-01 9.716952e-01 8.501671e-01
#>  [976] 2.234522e-01 7.960556e-01 4.740094e-01 4.814441e-01 6.854007e-01
#>  [981] 4.319057e-01 6.990788e-01 7.242594e-01 6.065816e-01 5.937351e-01
#>  [986] 3.409753e-01 7.929968e-01 5.444761e-02 7.359689e-01 7.302311e-01
#>  [991] 7.079751e-01 4.145144e-01 8.625015e-01 7.574569e-01 9.362680e-01
#>  [996] 5.270604e-01 6.478233e-01 1.839741e-04 5.798301e-01 2.044770e-01
#> [1001] 6.131697e-02 6.323071e-01 7.865934e-01 1.479641e-01 6.980816e-01
#> [1006] 7.840055e-01 5.847817e-02 5.390996e-01 5.814088e-01 5.732013e-01
#> [1011] 5.343370e-01 5.055327e-01 7.363840e-01 1.321850e-02 1.026463e-01
#> [1016] 8.584190e-01 4.997867e-02 2.573725e-01 6.732047e-01 6.739704e-01
#> [1021] 6.013213e-01 7.092178e-01 2.894415e-01 9.524205e-01 3.633874e-02
#> [1026] 9.328628e-01 7.288437e-01 4.232107e-02 6.437435e-01 4.606848e-01
#> [1031] 6.215071e-01 6.484520e-01 4.877567e-01 9.978976e-06 1.630421e-01
#> [1036] 5.450611e-01 1.472349e-01 1.196075e-01 1.061197e-01 5.833711e-01
#> [1041] 2.279277e-01 1.525488e-01 3.481648e-02 2.745511e-02 2.774369e-01
#> [1046] 5.239409e-01 6.385466e-01 9.517531e-02 5.615790e-01 1.460982e-01
#> [1051] 3.487266e-01 6.032475e-01 6.356615e-01 2.448774e-01 8.151241e-01
#> [1056] 7.472131e-01 7.281503e-01 8.674824e-01 4.112123e-01 4.796651e-01
#> [1061] 9.933613e-01 6.113651e-01 4.082479e-02 6.341450e-01 5.137795e-01
#> [1066] 9.441377e-01 2.215586e-01 7.506038e-01 5.136106e-01 2.321379e-01
#> [1071] 6.723355e-01 2.875934e-01 7.365840e-01 2.497677e-01 7.814087e-01
#> [1076] 5.899630e-01 1.276781e-01 5.787909e-01 3.636630e-01 3.420860e-01
#> [1081] 2.224319e-01 4.200601e-01 5.865269e-01 5.287036e-01 1.367006e-01
#> [1086] 4.837839e-01 2.790930e-01 3.042853e-01 6.854960e-01 7.844307e-01
#> [1091] 4.417747e-01 2.144410e-01 2.588829e-03 9.989997e-01 5.492882e-01
#> [1096] 2.341236e-01 2.645598e-01 2.771484e-01 4.771301e-01 9.953875e-01
#> [1101] 3.175513e-01 6.988095e-01 7.863369e-01 8.530413e-01 1.004617e-01
#> [1106] 2.076184e-01 3.180221e-01 3.002903e-01 3.307687e-01 6.165396e-01
#> [1111] 9.708517e-01 5.322649e-01 4.175468e-01 3.917172e-01 4.325122e-01
#> [1116] 8.518938e-01 4.297105e-01 7.376535e-01 3.407985e-01 5.537051e-01
#> [1121] 9.123923e-01 7.757075e-01 1.134731e-02 8.016511e-01 4.601832e-02
#> [1126] 5.060363e-01 2.079651e-01 8.552558e-01 1.195343e-01 2.510199e-01
#> [1131] 6.128421e-01 9.733815e-01 4.704637e-01 2.794488e-01 2.539861e-01
#> [1136] 4.424814e-01 3.829655e-01 5.389496e-01 9.101156e-01 2.402350e-02
#> [1141] 3.288329e-01 4.937795e-01 6.231806e-01 5.953893e-01 1.541625e-02
#> [1146] 9.133725e-01 8.240421e-01 6.570917e-01 1.440633e-01 1.253925e-02
#> [1151] 3.611406e-01 3.191929e-01 1.536041e-01 7.009747e-01 4.073813e-02
#> [1156] 5.012643e-01 1.485339e-01 1.453537e-01 7.100337e-01 4.863251e-01
#> [1161] 7.797377e-01 8.972185e-01 2.666217e-01 3.707083e-01 2.397798e-01
#> [1166] 9.070625e-01 5.765999e-01 9.947320e-04 2.320324e-01 4.743144e-01
#> [1171] 6.317229e-01 8.467173e-01 1.900928e-01 5.802671e-01 1.654541e-01
#> [1176] 8.657014e-01 2.736297e-01 5.511221e-02 1.800176e-01 6.580911e-01
#> [1181] 3.802181e-01 9.787469e-01 7.895271e-01 7.018761e-01 4.930704e-01
#> [1186] 5.489918e-01 4.136347e-01 7.834444e-01 5.492353e-01 6.961349e-01
#> [1191] 8.394399e-01 6.494153e-01 1.350972e-01 5.849692e-01 6.843793e-02
#> [1196] 9.119546e-01 1.336085e-01 7.654901e-01 5.460427e-04 9.645234e-01
#> [1201] 3.296774e-01 3.578427e-01 6.407573e-01 5.348217e-01 6.498206e-01
#> [1206] 6.392627e-01 6.155894e-01 5.726659e-01 6.058128e-01 2.636643e-01
#> [1211] 5.282586e-01 6.561365e-01 2.548537e-01 8.336645e-01 8.764472e-01
#> [1216] 6.953103e-01 9.124316e-01 8.115682e-01 5.297176e-01 6.476334e-01
#> [1221] 7.308857e-01 6.104571e-01 1.723419e-01 5.621151e-01 7.102242e-01
#> [1226] 9.094733e-01 5.259933e-01 6.998131e-02 5.941024e-01 1.437161e-01
#> [1231] 2.234554e-02 4.582523e-01 5.049988e-02 7.370357e-01 9.439987e-02
#> [1236] 7.483534e-01 1.902608e-01 5.395542e-01 6.404502e-01 3.520811e-01
#> [1241] 2.165731e-01 9.968339e-01 6.533055e-01 4.259191e-01 6.686194e-02
#> [1246] 9.082509e-02 5.908608e-01 3.898585e-01 6.196326e-01 3.327699e-01
#> [1251] 6.812416e-01 4.354438e-01 5.147339e-01 5.899737e-01 9.431568e-01
#> [1256] 4.972405e-01 9.453412e-01 5.279904e-01 5.179368e-01 4.760335e-01
#> [1261] 3.777444e-01 7.811007e-01 2.830383e-01 4.108834e-01 5.222929e-01
#> [1266] 7.525192e-01 5.151110e-01 7.244818e-01 9.803420e-01 3.690573e-01
#> [1271] 4.077826e-01 6.939638e-01 5.148259e-01 5.641640e-01 1.132963e-01
#> [1276] 9.977271e-01 8.109103e-01 2.086745e-01 5.264247e-01 9.977442e-01
#> [1281] 1.667127e-02 4.202332e-01 4.704066e-01 1.647978e-01 9.968035e-01
#> [1286] 7.285440e-01 9.307384e-01 5.708360e-01 5.172131e-01 3.174449e-01
#> [1291] 7.546071e-01 5.561122e-01 3.289950e-01 5.739173e-01 8.673324e-01
#> [1296] 9.897459e-01 5.352231e-01 9.995226e-01 4.859332e-01 5.497801e-01
#> [1301] 4.904069e-01 5.028268e-01 1.450044e-01 7.132298e-01 2.338136e-01
#> [1306] 3.417817e-01 3.876989e-01 3.902007e-01 7.077205e-01 6.662023e-01
#> [1311] 2.870865e-01 9.770520e-01 9.255915e-01 5.266458e-01 2.681689e-01
#> [1316] 5.528317e-01 3.064378e-01 7.418612e-01 5.746926e-01 3.813103e-01
#> [1321] 4.595744e-01 7.693796e-02 3.602099e-03 3.314173e-01 9.681951e-01
#> [1326] 9.219397e-01 9.661995e-03 7.047082e-01 4.729542e-01 8.861677e-01
#> [1331] 9.949900e-01 5.756855e-01 9.130566e-04 6.232607e-01 5.234629e-01
#> [1336] 2.978732e-01 8.300812e-01 3.842922e-01 8.551052e-01 7.353335e-01
#> [1341] 3.098419e-01 6.646053e-01 1.518113e-01 5.464905e-01 2.304573e-01
#> [1346] 9.745618e-01 6.794120e-01 3.009589e-01 1.835807e-01 2.593853e-01
#> [1351] 5.312646e-01 3.564579e-01 6.181097e-01 6.275430e-01 8.605840e-01
#> [1356] 5.133349e-01 2.269185e-02 9.828029e-02 6.450927e-01 2.934726e-01
#> [1361] 2.668214e-01 8.617854e-01 1.903812e-01 5.467492e-01 4.842068e-01
#> [1366] 6.732276e-01 3.696574e-01 8.998469e-01 8.332951e-01 4.257767e-01
#> [1371] 2.306594e-01 1.585202e-01 8.821272e-01 5.577664e-01 2.484513e-01
#> [1376] 5.685280e-01 8.121377e-01 9.410109e-02 2.337773e-01 2.771860e-01
#> [1381] 9.995703e-01 4.295585e-01 1.676488e-02 8.250009e-01 5.901499e-01
#> [1386] 4.208111e-01 9.993874e-01 2.288132e-01 4.465796e-01 2.129054e-03
#> [1391] 2.067362e-01 9.849378e-01 8.988156e-01 1.008570e-01 3.324391e-02
#> [1396] 5.842542e-01 5.687899e-01 7.109624e-01 3.916386e-01 9.850554e-02
#> [1401] 4.796001e-01 6.737459e-01 6.976638e-01 2.461717e-01 4.539836e-01
#> [1406] 6.003839e-01 6.555489e-01 4.391596e-01 4.309909e-01 4.590733e-03
#> [1411] 1.275848e-02 2.069089e-01 5.491119e-01 9.544223e-01 5.897076e-02
#> [1416] 5.906477e-01 2.686600e-01 5.216788e-01 6.278538e-01 6.227296e-01
#> [1421] 7.223309e-01 8.414838e-01 4.256734e-01 1.050389e-03 7.847142e-01
#> [1426] 2.577009e-01 6.992939e-01 4.674525e-01 8.690383e-01 1.408397e-01
#> [1431] 6.747608e-01 1.369545e-02 6.841913e-01 5.242086e-01 6.821707e-01
#> [1436] 5.860246e-02 4.475625e-01 4.874239e-01 5.327015e-01 9.164135e-02
#> [1441] 4.935063e-01 4.854192e-01 4.950523e-01 9.479049e-01 9.361872e-01
#> [1446] 4.199128e-01 7.892321e-01 5.844838e-02 7.785713e-01 6.004307e-01
#> [1451] 9.445488e-01 1.052228e-02 5.998176e-01 7.353462e-01 3.388574e-01
#> [1456] 3.795280e-01 8.667023e-01 9.204133e-01 5.708362e-01 5.885191e-01
#> [1461] 3.130555e-03 6.557072e-01 8.043957e-01 8.326593e-01 1.689500e-01
#> [1466] 4.191427e-02 6.654033e-01 9.882046e-01 7.592022e-01 8.112903e-01
#> [1471] 3.439061e-01 1.610789e-01 8.666471e-01 2.031419e-01 7.866233e-01
#> [1476] 5.831468e-01 7.790083e-03 4.929816e-01 9.506979e-01 9.501340e-01
#> [1481] 7.334860e-01 4.752531e-01 7.071570e-02 6.147970e-01 6.283280e-01
#> [1486] 6.589783e-01 7.596325e-01 4.009038e-01 2.346058e-01 7.187654e-01
#> [1491] 6.222667e-01 9.798723e-01 4.738730e-01 8.455414e-03 3.727805e-01
#> [1496] 4.118097e-01 9.916952e-01 2.252968e-01 3.391240e-01 6.140360e-01
#> [1501] 1.596798e-01 6.327046e-01 2.783617e-01 9.777698e-01 6.357478e-01
#> [1506] 6.455359e-01 4.913595e-01 9.919015e-01 4.196130e-01 7.328468e-01
#> [1511] 5.847946e-01 5.816470e-01 5.086931e-01 7.772806e-01 5.303061e-01
#> [1516] 6.425071e-01 2.374804e-01 5.271806e-01 9.677957e-02 5.293094e-01
#> [1521] 7.111186e-01 8.422106e-01 7.593508e-01 2.963860e-01 5.786188e-01
#> [1526] 1.712765e-01 5.670507e-01 6.000422e-01 9.656145e-01 2.518458e-01
#> [1531] 4.878762e-01 6.210219e-01 1.884914e-01 5.808606e-01 3.117062e-01
#> [1536] 9.565780e-01 1.559619e-02 3.857087e-01 9.706592e-01 6.046307e-01
#> [1541] 2.197492e-01 4.347638e-01 4.712997e-01 5.910085e-01 6.257280e-01
#> [1546] 5.075163e-01 9.766908e-01 1.161802e-01 2.850652e-01 3.464883e-01
#> [1551] 7.275777e-01 5.748951e-01 6.509088e-01 4.428615e-01 3.869068e-01
#> [1556] 5.081354e-01 4.199475e-01 4.211777e-01 1.177133e-01 6.781556e-01
#> [1561] 7.827020e-01 3.828734e-01 3.299453e-01 2.330421e-01 3.587822e-01
#> [1566] 2.548504e-01 1.136597e-01 1.325975e-01 4.352043e-01 7.195918e-01
#> [1571] 3.206768e-01 3.059745e-01 5.436882e-01 9.983418e-01 6.335386e-01
#> [1576] 5.606764e-01 4.539807e-01 9.361687e-01 9.595573e-01 7.952185e-02
#> [1581] 2.650461e-01 5.988888e-01 9.615660e-01 2.483964e-01 9.744439e-01
#> [1586] 6.559892e-01 4.127007e-01 2.351443e-01 4.723982e-01 5.698518e-01
#> [1591] 1.417783e-01 6.175851e-02 5.279490e-01 3.613598e-01 5.092401e-01
#> [1596] 9.286889e-01 5.768206e-01 6.732502e-01 3.847434e-01 8.545808e-01
#> [1601] 5.727247e-01 6.371998e-01 4.269194e-01 5.185764e-01 8.099305e-01
#> [1606] 8.955387e-01 4.517871e-01 4.070390e-01 1.062187e-01 9.276757e-01
#> [1611] 3.051159e-01 3.741310e-01 5.111835e-01 2.127126e-02 5.314329e-01
#> [1616] 9.465652e-01 1.896074e-01 8.212704e-01 3.905015e-01 2.642736e-02
#> [1621] 9.693233e-01 6.458365e-01 4.415667e-01 6.134717e-01 2.975209e-01
#> [1626] 6.852228e-01 1.671462e-02 7.826871e-01 7.319493e-01 6.138181e-01
#> [1631] 9.944238e-01 4.838609e-01 4.242751e-02 6.520197e-01 7.879134e-01
#> [1636] 5.091839e-01 9.103646e-01 9.945322e-01 9.989007e-01 6.969048e-01
#> [1641] 4.775657e-01 3.573385e-01 9.553975e-01 6.931314e-01 6.118806e-01
#> [1646] 7.625415e-01 4.771316e-01 1.688261e-01 1.356178e-02 6.206565e-01
#> [1651] 8.770972e-02 2.501906e-01 9.334125e-01 5.441661e-01 2.511751e-01
#> [1656] 3.417813e-01 5.746998e-01 8.869532e-01 5.206274e-03 4.374536e-01
#> [1661] 5.255866e-01 6.747227e-01 7.274760e-01 8.763817e-01 3.404405e-01
#> [1666] 3.811784e-01 3.754904e-01 3.910091e-01 3.388121e-01 6.011923e-01
#> [1671] 6.143306e-01 1.402799e-01 2.479189e-01 1.130392e-01 9.998099e-01
#> [1676] 4.345770e-01 9.517984e-02 7.217396e-01 4.948191e-01 1.614617e-01
#> [1681] 8.275871e-01 1.520523e-01 8.800903e-01 6.264619e-01 6.634759e-01
#> [1686] 7.910772e-01 2.683169e-01 8.383818e-01 5.691127e-01 9.720919e-01
#> [1691] 4.799091e-01 4.556106e-01 5.122294e-01 8.297605e-01 1.862962e-01
#> [1696] 9.882299e-01 4.802663e-01 5.620263e-01 3.207568e-01 3.897623e-01
#> [1701] 4.902030e-01 4.597808e-01 2.148257e-01 7.953964e-01 5.653577e-01
#> [1706] 9.935444e-01 9.937705e-01 6.310983e-03 4.258335e-01 2.427561e-01
#> [1711] 3.777704e-01 1.982468e-01 7.674848e-01 6.423465e-01 4.327719e-01
#> [1716] 3.407863e-01 3.064204e-01 1.117669e-01 7.903382e-02 5.718648e-01
#> [1721] 1.749660e-02 2.184884e-01 2.039105e-02 7.351064e-01 9.392714e-01
#> [1726] 5.496263e-01 9.396409e-01 4.518385e-01 6.407353e-01 6.832618e-01
#> [1731] 2.156719e-02 4.547112e-02 1.032958e-01 1.010101e-01 6.012912e-01
#> [1736] 7.393891e-01 8.340310e-01 5.694919e-01 8.863026e-01 7.476427e-01
#> [1741] 9.162117e-01 3.021176e-01 6.822338e-01 4.973812e-01 3.913400e-01
#> [1746] 8.695274e-01 5.716337e-01 4.054399e-03 6.787327e-01 3.019045e-01
#> [1751] 3.411203e-01 3.209316e-01 6.721530e-01 2.390198e-01 8.787920e-01
#> [1756] 5.817333e-01 2.982527e-01 7.364411e-01 8.098504e-01 4.799467e-01
#> [1761] 6.387274e-01 2.064130e-01 6.722309e-01 2.593375e-01 3.851455e-01
#> [1766] 6.793444e-01 5.431186e-01 5.629742e-01 3.344066e-03 4.314044e-01
#> [1771] 5.860999e-01 7.045471e-01 4.360370e-01 8.791629e-01 3.140622e-01
#> [1776] 4.180700e-01 7.489595e-01 2.491531e-02 7.967895e-01 6.998083e-01
#> [1781] 9.719999e-01 9.982542e-01 5.017160e-01 3.282477e-01 2.954772e-01
#> [1786] 7.716683e-01 3.966822e-01 4.611353e-01 4.389913e-01 3.555861e-02
#> [1791] 2.794121e-01 7.275591e-01 3.710198e-02 2.000496e-01 2.789230e-01
#> [1796] 6.789118e-01 6.515143e-01 5.940844e-01 4.332465e-01 7.579573e-01
#> [1801] 5.750565e-01 5.233543e-01 3.929975e-01 2.594591e-01 2.366918e-03
#> [1806] 4.735079e-01 6.031430e-01 6.372081e-01 8.946001e-02 7.066084e-01
#> [1811] 7.148262e-01 8.838530e-01 4.825109e-01 4.175559e-01 5.167493e-01
#> [1816] 6.983707e-01 5.150429e-01 4.234627e-03 6.835199e-01 2.085124e-01
#> [1821] 5.916931e-01 1.830303e-01 7.212672e-03 7.204078e-01 3.734913e-01
#> [1826] 5.094491e-01 4.507095e-01 3.793425e-01 8.586045e-02 7.738846e-01
#> [1831] 6.023812e-01 2.240659e-01 5.277852e-01 3.815685e-01 1.690874e-01
#> [1836] 1.507372e-01 4.189307e-01 4.146400e-01 8.480710e-01 4.742373e-02
#> [1841] 5.923090e-01 3.558568e-01 4.176485e-01 6.044230e-01 1.408504e-01
#> [1846] 2.268343e-01 1.127294e-01 2.210104e-01 5.604168e-01 3.823551e-02
#> [1851] 9.531479e-01 6.225686e-01 1.055704e-01 4.036839e-01 3.062785e-01
#> [1856] 6.781945e-01 5.023901e-01 7.397493e-01 6.602551e-01 1.609746e-01
#> [1861] 7.788966e-01 6.957264e-01 5.959645e-01 6.768573e-01 7.342269e-03
#> [1866] 6.782820e-01 5.920132e-01 5.399067e-01 2.091592e-01 6.146463e-01
#> [1871] 6.186586e-01 7.939086e-01 7.323228e-01 5.236885e-01 8.272626e-01
#> [1876] 1.466087e-01 5.266806e-01 2.823292e-01 3.045121e-01 6.686728e-01
#> [1881] 3.380545e-01 6.499160e-01 1.343901e-01 4.125626e-01 2.268452e-01
#> [1886] 5.077862e-01 1.076238e-01 2.488840e-01 1.330912e-01 5.385482e-01
#> [1891] 7.963307e-01 4.301725e-01 6.171397e-01 2.052950e-03 6.174852e-01
#> [1896] 1.903783e-01 9.862635e-01 8.991319e-01 1.877129e-01 4.707806e-02
#> [1901] 6.487581e-01 5.953579e-01 5.997223e-01 2.211870e-01 7.723163e-01
#> [1906] 6.765108e-01 6.257692e-01 2.958351e-01 5.929265e-01 5.345449e-01
#> [1911] 4.427641e-01 3.365980e-01 1.252136e-01 1.732857e-01 4.242970e-01
#> [1916] 5.217419e-01 1.740722e-01 2.627098e-01 4.946286e-01 3.284404e-01
#> [1921] 2.033060e-02 1.513034e-01 3.017813e-01 5.268998e-02 4.456686e-01
#> [1926] 4.930353e-01 9.408174e-01 4.893942e-01 5.788809e-01 2.823301e-01
#> [1931] 5.218593e-01 8.169109e-01 2.494612e-01 1.852622e-01 3.436880e-01
#> [1936] 3.150536e-01 7.406442e-01 4.413533e-01 9.912757e-01 4.860385e-01
#> [1941] 7.240070e-01 5.114848e-01 6.271777e-01 4.877836e-01 4.850605e-02
#> [1946] 6.113550e-01 3.314411e-01 9.123951e-02 9.840932e-01 7.795874e-01
#> [1951] 6.484160e-01 1.135360e-01 9.699098e-01 5.671421e-01 4.264446e-01
#> [1956] 3.120284e-01 3.138598e-01 4.212018e-01 3.757032e-01 5.576084e-01
#> [1961] 7.341078e-01 5.245742e-01 6.470752e-01 5.993320e-01 6.921317e-01
#> [1966] 5.343384e-01 3.385299e-01 5.293995e-01 4.242423e-01 8.601112e-01
#> [1971] 2.166673e-02 6.763705e-01 4.147982e-01 4.521916e-01 4.613753e-01
#> [1976] 4.940971e-01 7.690830e-01 5.384153e-01 8.519834e-01 7.491064e-01
#> [1981] 4.524284e-01 4.738410e-01 4.818030e-01 6.554952e-01 5.644121e-01
#> [1986] 7.057145e-01 4.614571e-01 4.019925e-01 3.476020e-01 4.817820e-01
#> [1991] 8.662857e-01 4.901267e-01 6.770929e-01 8.565970e-01 5.469539e-01
#> [1996] 4.842959e-01 3.267069e-01 3.968082e-01 5.267576e-02 6.669564e-02